refactored: books dtos is moved to author package -set input and update dto to final -added validation to input and update
